Perform explicit trusted check as auth
Begin setting an Authenticated boolean that will report if the NarwhalMessage likely came from an authenticated, trusted user.

This is determined by performing a lookup of the user in all channels the client is in (thus, the user must be in a channel that the bot is). If the message is from a user in a channel, we look up their permissions for that channel and mark their authenticated state as:

1. If they are active
2. If they are trusted (e.g. not a normal user, should be voice, half op, op, etc.)

If it came from a direct message, we require an explicit full user / ident + host in the admin list to perform the action.