Nuke all that invite work.
We are unable to query channel or user details from the IRC server, the details are purely internal girc state and thus not useful when the bot isn't in the server. As a result of not being able to query those details, we can't determine who is authorized outside of our pre-defined admins in the config.

So just accept that things are the way they are, be sad and revert.