Separate out Join handling, add kick handling. Tweak log output.
Separated out the join-specific logic (we only really care about autokick here) so we don't have to concern ourselves with the rest of the plugins. Added a KICK handler, right now just to log the NarwhalMessage output so we can see what sort of info we are provided.

Instead of overly verbose trunk info logging for every action, log only messages for now with our new PrintPrettyMessage func. We can spruce this up in the future to log leave and join if we want.

Nuke old structs that haven't been used since we switched to sauron.